// angular-ui/src/app/services/models/product-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ProductResponse {
  description?: string;
  expiryDate?: string;
  image?: string;
  manufacturer?: string;
  name?: string;
  price?: number;
  productId?: number;
  stockQuantity?: number;
  type?: 'SYRUP' | 'TABLETS' | 'OINTMENTS';
}